v2 WIP Initial cleanup
